# application/services/custom_services/ticketing/src/swagger/swagger.yaml
---
openapi: 3.0.0
info:
   version: 1.3.0
   title: ticketing open api specification
   description: ticketing API
   termsOfService: http://api_url/terms/
   contact:
      name: ContactName
      email: hello@wolox.co
      url: https://www.wolox.com.ar/
   license:
      name: Apache 2.0
      url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
   - url: http://localhost:8007
     description: ticketing
tags:
- name: travel
paths:
   "/travel":
      post:
         tags:
         -  travel
         description: GpCreate
         requestBody:
            content:
               application/json:
                  schema:
                     $ref: "#/components/schemas/travel"
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
      put:
         tags:
         -  travel
         description: GpUpdate
         requestBody:
            content:
               application/json:
                  schema:
                     $ref: "#/components/schemas/travel"
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
      get:
         tags:
         -  travel
         description: GpGetAllValues
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
   "/travel/{id}":
      delete:
         tags:
         -  travel
         description: GpDelete
         parameters:
         -  in: "path"
            description: GpDelete
            name: id
            required: true
            schema:
               type: "string"
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
   "/travel/get/search":
      get:
         tags:
         -  travel
         description: GpSearch
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
   "/travel/get/update":
      put:
         tags:
         -  travel
         description: GpSearchForUpdate
         requestBody:
            content:
               application/json:
                  schema:
                     $ref: "#/components/schemas/travel"
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
components:
   schemas:
      travel:
         type: object
         properties:
            name:
               type: "string"
               description: Description
            age:
               type: "number"
               description: Description
            email:
               type: "string"
               description: Description
            phone:
               type: "number"
               description: Description